refactor: major package 리펙토링
