+++ src/main/java/com/capstone/wanf/course/domain/repo/MajorRepository.java
@@ -1,6 +1,6 @@
-package com.capstone.wanf.major.domain.entity.repo;
+package com.capstone.wanf.course.domain.repo;
 
-import com.capstone.wanf.major.domain.entity.Major;
+import com.capstone.wanf.course.domain.entity.Major;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MajorRepository extends JpaRepository<Major, Long> {
